set up proto generation

+++ app/src/main/java/com/strobel/emercast/db/models/Authority.kt
@@ -2,11 +2,11 @@ package com.strobel.emercast.db.models
 
 data class Authority (
     val id: String,
-    val created: Int,
+    val created: Long,
     val createdBy: String,
     val publicName: String,
-    val validUntil: Int,
+    val validUntil: Long,
     val publicKeyBase64: String,
 
-    val revokedAfter: Int?
+    val revokedAfter: Long?
 )
