+++ app/src/main/java/com/strobel/emercast/db/models/JurisdictionMarker.kt
@@ -2,7 +2,7 @@ package com.strobel.emercast.db.models
 
 data class JurisdictionMarker (
     val authorityId: String,
-    val authorityCreated: Int,
+    val authorityCreated: Long,
     val latitude: Float,
     val longitude: Float,
     val kind: String,
